requirements.txt
django>=4.0,<4.2
djangorestframework>=3.12.4,<3.15
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.16.0,<0.23.0